@@ -1,17 +1,17 @@
 name: flutter_openvpn
 description: Flutter OpenVPN uses OpenVpn Protocol to connect with ovpn file configurations.
-version: 0.2.0
+version: 0.3.0
 homepage: https://gitlab.com/topfreelancerdeveloper/flutter_openvpn.git
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  intl: ^0.16.1
-  shared_preferences: ^0.5.8
+  intl: ^0.17.0
+  shared_preferences: ^2.0.8
 
 dev_dependencies:
   flutter_test:
